refactor. Layout 코드 분리
